+++ evaluator.py
@@ -9,9 +9,9 @@ term = Terminal()
 
 def main():
     if len(sys.argv) >= 2:
-        if sys.argv[1] >= "2.0.5":
+        if sys.argv[1] >= "2.1.0":
             # Initialize with splash screen
-            ekernel.splashScreen("ProcyonCLS Evaluator", "Version 2.0.5")
+            ekernel.splashScreen("ProcyonCLS Evaluator", "Version 2.1.0")
             
             while True:
                 # Display interface